Added multiple dvl sensors and modified top-level launch files
