+++ models/dave_object_models/launch/upload_object.launch.py
@@ -8,19 +8,67 @@ from launch_ros.actions import Node
 
 
 def generate_launch_description():
-    object_name = LaunchConfiguration("object_name")
-    use_sim = LaunchConfiguration("use_sim")
+    gui = LaunchConfiguration("gui")
+    use_sim_time = LaunchConfiguration("use_sim_time")
+    namespace = LaunchConfiguration("namespace")
+    x = LaunchConfiguration("x")
+    y = LaunchConfiguration("y")
+    z = LaunchConfiguration("z")
+    roll = LaunchConfiguration("roll")
+    pitch = LaunchConfiguration("pitch")
+    yaw = LaunchConfiguration("yaw")
+    use_ned_frame = LaunchConfiguration("use_ned_frame")
 
     args = [
         DeclareLaunchArgument(
-            "object_name",
-            default_value="model",
-            description="Name of the object model to load",
-        ),
-        DeclareLaunchArgument(
-            "use_sim",
+            "gui",
             default_value="true",
             description="Flag to indicate whether to use simulation",
+        ),
+        DeclareLaunchArgument(
+            "use_sim_time",
+            default_value="true",
+            description="Flag to indicate whether to use sim time",
+        ),
+        DeclareLaunchArgument(
+            "namespace",
+            default_value="",
+            description="Namespace",
+        ),
+        DeclareLaunchArgument(
+            "x",
+            default_value="0",
+            description="Initial x position",
+        ),
+        DeclareLaunchArgument(
+            "y",
+            default_value="0",
+            description="Initial y position",
+        ),
+        DeclareLaunchArgument(
+            "z",
+            default_value="0.0",
+            description="Initial z position",
+        ),
+        DeclareLaunchArgument(
+            "roll",
+            default_value="0.0",
+            description="Initial roll",
+        ),
+        DeclareLaunchArgument(
+            "pitch",
+            default_value="0.0",
+            description="Initial pitch",
+        ),
+        DeclareLaunchArgument(
+            "yaw",
+            default_value="0.0",
+            description="Initial yaw",
+        ),
+        DeclareLaunchArgument(
+            "use_ned_frame",
+            default_value="false",
+            description="Use North-East-Down frame",
         ),
     ]
 
@@ -28,25 +76,61 @@ def generate_launch_description():
         [
             FindPackageShare("dave_object_models"),
             "description",
-            object_name,
+            namespace,
             "model.sdf",
         ]
+    )
+
+    tf2_spawner = Node(
+        package="tf2_ros",
+        executable="static_transform_publisher",
+        name="world_to_world_ned",
+        arguments=[
+            "--roll",
+            "1.57",
+            "--yaw",
+            "3.14",
+            "--frame_id",
+            "world",
+            "--child_frame_id",
+            "world_ned",
+        ],
+        output="both",
+        condition=IfCondition(use_ned_frame),
+        parameters=[{"use_sim_time": use_sim_time}],
     )
 
     gz_spawner = Node(
         package="ros_gz_sim",
         executable="create",
-        arguments=["-name", object_name, "-file", description_file],
+        arguments=[
+            "-name",
+            namespace,
+            "-file",
+            description_file,
+            "-x",
+            x,
+            "-y",
+            y,
+            "-z",
+            z,
+            "-roll",
+            roll,
+            "-pitch",
+            pitch,
+            "-yaw",
+            yaw,
+        ],
         output="both",
-        condition=IfCondition(use_sim),
-        parameters=[{"use_sim_time": use_sim}],
+        condition=IfCondition(gui),
+        parameters=[{"use_sim_time": use_sim_time}],
     )
 
-    nodes = [gz_spawner]
+    nodes = [tf2_spawner, gz_spawner]
 
     event_handlers = [
         RegisterEventHandler(
-            OnProcessExit(target_action=gz_spawner, on_exit=LogInfo(msg="Model Uploaded"))
+            OnProcessExit(target_action=gz_spawner, on_exit=LogInfo(msg="Object Model Uploaded"))
         )
     ]
