+++ models/dave_sensor_models/description/nortek_dvl500_6000/model.config
@@ -1,0 +1,11 @@
+<?xml version="1.0" ?>
+<model>
+    <name>nortek_dvl500_6000</name>
+    <version>1.0</version>
+    <sdf version="1.7">model.sdf</sdf>
+    <author>
+        <name>Duane Davis</name>
+        <email>dtdavi1@nps.edu</email>
+    </author>
+    <description>Nortek DVL500-6000</description>
+</model>
