+++ models/dave_sensor_models/description/nortek_dvl500_6000/model.sdf
@@ -1,18 +1,18 @@
 <?xml version="1.0"?>
 <sdf version="1.7">
-  <model name="dvl500_300">
+  <model name="nortek_dvl500_6000">
     <link name="dvl500_base_link">
       <pose>0 0 0 0 0 0</pose>
       <inertial>
         <pose>0 0 0 0 0 0</pose>
-        <mass>3.5</mass>
+        <mass>3.1</mass>
         <inertia>
-          <ixx>0.0195872</ixx>
+          <ixx>0.0330184</ixx>
           <ixy>0</ixy>
           <ixz>0</ixz>
-          <iyy>0.0195872</iyy>
+          <iyy>0.0330184</iyy>
           <iyz>0</iyz>
-          <izz>0.0151357</izz>
+          <izz>0.0255146</izz>
         </inertia>
       </inertial>
       <velocity_decay/>
@@ -20,7 +20,7 @@
       <velocity_decay/>
 
       <sensor
-        name="nortek_dvl500_300" type="custom"
+        name="nortek_dvl500_6000" type="custom"
         gz:type="dvl">
         <pose>0 0 0 0 0 0</pose>
         <always_on>1</always_on>
@@ -92,13 +92,13 @@
           <mesh>
             <scale>1 1 1</scale>
             <!-- Mesh is actually for a Nortek DVL1000-4000m for now -->
-            <uri>model://meshes/nortek_dvl500_300/DVL500-300m.dae</uri>
+            <uri>model://meshes/nortek_dvl500_6000/DVL500-6000m.dae</uri>
           </mesh>
         </geometry>
         <transparency>0</transparency>
         <cast_shadows>1</cast_shadows>
       </visual>
-      <collision name="dvl500_base_link_collision">
+      <collision name="dvl500_base_link_fixed_joint_lump__dvl500_link_collision">
         <laser_retro>0</laser_retro>
         <max_contacts>10</max_contacts>
         <pose>0 0 0 3.14159 0 0</pose>
