+++ models/dave_sensor_models/description/sonardyne_syrinx600/model.config
@@ -1,0 +1,11 @@
+<?xml version="1.0" ?>
+<model>
+    <name>sonardyne_syrinx600</name>
+    <version>1.0</version>
+    <sdf version="1.6">model.sdf</sdf>
+    <author>
+        <name>Duane Davis</name>
+        <email>dtdavi1@nps.edu</email>
+    </author>
+    <description>Sonardyne Syrinx 600 model utilizing the WHOI DSL Gazebo environment DVL plugin</description>
+</model>
