+++ models/dave_sensor_models/description/teledyne_explorer4000/model.config
@@ -1,0 +1,11 @@
+<?xml version="1.0" ?>
+<model>
+    <name>teledyne_explorer4000</name>
+    <version>1.0</version>
+    <sdf version="1.6">model.sdf</sdf>
+    <author>
+        <name>Duane Davis</name>
+        <email>dtdavi1@nps.edu</email>
+    </author>
+    <description>Teledyne Explorer 4000 model utilizing the WHOI DSL Gazebo environment DVL plugin</description>
+</model>
