+++ models/dave_sensor_models/description/teledyne_explorer4000/model.sdf
@@ -1,18 +1,18 @@
 <?xml version="1.0"?>
-<sdf version="1.7">
-  <model name="dvl500_300">
-    <link name="dvl500_base_link">
+<sdf version="1.6">
+  <model name="teledyne_explorer4000">
+    <link name="explorer4000_base_link">
       <pose>0 0 0 0 0 0</pose>
       <inertial>
         <pose>0 0 0 0 0 0</pose>
-        <mass>3.5</mass>
+        <mass>8.6</mass>
         <inertia>
-          <ixx>0.0195872</ixx>
+          <ixx>0.0219827</ixx>
           <ixy>0</ixy>
           <ixz>0</ixz>
-          <iyy>0.0195872</iyy>
+          <iyy>0.0957988</iyy>
           <iyz>0</iyz>
-          <izz>0.0151357</izz>
+          <izz>0.0957988</izz>
         </inertia>
       </inertial>
       <velocity_decay/>
@@ -20,34 +20,34 @@
       <velocity_decay/>
 
       <sensor
-        name="nortek_dvl500_300" type="custom"
+        name="teledyne_explorer4000" type="custom"
         gz:type="dvl">
         <pose>0 0 0 0 0 0</pose>
         <always_on>1</always_on>
-        <update_rate>8</update_rate>
+        <update_rate>12</update_rate>
         <topic>/dvl/velocity</topic>
         <gz:dvl>
           <type>phased_array</type>
           <arrangement degrees="true">
             <beam id="1">
-              <aperture>3</aperture>
+              <aperture>3.8</aperture>
               <rotation>-135</rotation>
-              <tilt>25</tilt>
+              <tilt>30</tilt>
             </beam>
             <beam>
-              <aperture>3</aperture>
+              <aperture>3.8</aperture>
               <rotation>135</rotation>
-              <tilt>25</tilt>
+              <tilt>30</tilt>
             </beam>
             <beam>
-              <aperture>3</aperture>
+              <aperture>3.8</aperture>
               <rotation>45</rotation>
-              <tilt>25</tilt>
+              <tilt>30</tilt>
             </beam>
             <beam>
-              <aperture>3</aperture>
+              <aperture>3.8</aperture>
               <rotation>-45</rotation>
-              <tilt>25</tilt>
+              <tilt>30</tilt>
             </beam>
           </arrangement>
           <tracking>
@@ -64,20 +64,20 @@
               </boundaries>
               <bins>10</bins>
               <noise type="gaussian">
-                <stddev>0.015</stddev>
+                <stddev>0.025</stddev>
               </noise>
               <visualize>false</visualize>
             </water_mass_mode> -->
             <bottom_mode>
               <when>best</when>
               <noise type="gaussian">
-                <stddev>0.005</stddev>
+                <stddev>0.019</stddev>
               </noise>
               <visualize>true</visualize>
             </bottom_mode>
           </tracking>
-          <maximum_range>200.</maximum_range>
-          <minimum_range>0.3</minimum_range>
+          <maximum_range>66.</maximum_range>
+          <minimum_range>0.5</minimum_range>
           <!-- ENU to SFM -->
           <reference_frame>0 0 0 0 0 -1.570796</reference_frame>
         </gz:dvl>
@@ -86,26 +86,26 @@
       <self_collide>0</self_collide>
       <enable_wind>0</enable_wind>
       <kinematic>0</kinematic>
-      <visual name="dvl500_base_link_visual">
-        <pose>0 0 0 0 0 3.14159</pose>
+      <visual name="explorer4000_base_link_visual">
+        <pose>0 0 0 1.5708 0 1.5708</pose>
         <geometry>
           <mesh>
             <scale>1 1 1</scale>
-            <!-- Mesh is actually for a Nortek DVL1000-4000m for now -->
-            <uri>model://meshes/nortek_dvl500_300/DVL500-300m.dae</uri>
+            <!-- Mesh is actually for a Teledyne Explorer4000 for now -->
+            <uri>model://meshes/teledyne_explorer4000/MarineExplorer.dae</uri>
           </mesh>
         </geometry>
         <transparency>0</transparency>
         <cast_shadows>1</cast_shadows>
       </visual>
-      <collision name="dvl500_base_link_collision">
+      <collision name="explorer4000_base_link_collision">
         <laser_retro>0</laser_retro>
         <max_contacts>10</max_contacts>
-        <pose>0 0 0 3.14159 0 0</pose>
+        <pose>0 0 0 3.14159 -1.57079 0</pose>
         <geometry>
           <cylinder>
-            <radius>0.093</radius>
-            <length>0.203</length>
+            <radius>0.0715</radius>
+            <length>0.344</length>
           </cylinder>
         </geometry>
         <surface>
