+++ models/dave_sensor_models/description/teledyne_whn/model.config
@@ -1,0 +1,11 @@
+<?xml version="1.0" ?>
+<model>
+    <name>teledyne_whn_dsl</name>
+    <version>1.0</version>
+    <sdf version="1.6">model.sdf</sdf>
+    <author>
+        <name>Duane Davis</name>
+        <email>dtdavi1@nps.edu</email>
+    </author>
+    <description>Teledyne WHN DVL model compatible with the Woods Hole Oceanographic Institute Deep Submergence Laboratory Gazebo environment plugin</description>
+</model>
