+++ models/dave_sensor_models/description/teledyne_whn/model.sdf
@@ -1,50 +1,52 @@
 <?xml version="1.0"?>
+<!-- Teledyne Workhorse Navigator Doppler Velocity Logger (DVL) utilizing the
+     WHOI DSL Gazebo environment DVL plugin -->
 <sdf version="1.6">
-  <model name="teledyne_explorer1000">
-    <link name="explorer_base_link">
+  <model name="teledyne_whn_dsl">
+    <link name="whn_base_link">
       <pose>0 0 0 0 0 0</pose>
       <inertial>
         <pose>0 0 0 0 0 0</pose>
-        <mass>4.3</mass>
+        <mass>15.8</mass>
         <inertia>
-          <ixx>0.0082646</ixx>
+          <ixx>0.1287917645</ixx>
           <ixy>0</ixy>
           <ixz>0</ixz>
-          <iyy>0.0422145</iyy>
+          <iyy>0.1287917645</iyy>
           <iyz>0</iyz>
-          <izz>0.0422145</izz>
+          <izz>0.100162204</izz>
         </inertia>
       </inertial>
       <velocity_decay/>
-      <gravity>1</gravity>
+      <gravity>0</gravity>
 
       <sensor
-        name="teledyne_explorer1000" type="custom"
+        name="teledyne_whn_dsl" type="custom"
         gz:type="dvl">
         <pose>0 0 0 0 0 0</pose>
         <always_on>1</always_on>
-        <update_rate>12</update_rate>
+        <update_rate>7</update_rate>
         <topic>/dvl/velocity</topic>
         <gz:dvl>
           <type>phased_array</type>
           <arrangement degrees="true">
             <beam id="1">
-              <aperture>2.2</aperture>
+              <aperture>4</aperture>
               <rotation>-135</rotation>
               <tilt>30</tilt>
             </beam>
             <beam>
-              <aperture>2.2</aperture>
+              <aperture>4</aperture>
               <rotation>135</rotation>
               <tilt>30</tilt>
             </beam>
             <beam>
-              <aperture>2.2</aperture>
+              <aperture>4</aperture>
               <rotation>45</rotation>
               <tilt>30</tilt>
             </beam>
             <beam>
-              <aperture>2.2</aperture>
+              <aperture>4</aperture>
               <rotation>-45</rotation>
               <tilt>30</tilt>
             </beam>
@@ -63,20 +65,20 @@
               </boundaries>
               <bins>10</bins>
               <noise type="gaussian">
-                <stddev>0.048</stddev>
+                <stddev>0.0075</stddev>
               </noise>
               <visualize>false</visualize>
             </water_mass_mode> -->
             <bottom_mode>
               <when>best</when>
               <noise type="gaussian">
-                <stddev>0.018</stddev>
+                <stddev>0.005</stddev>
               </noise>
               <visualize>true</visualize>
             </bottom_mode>
           </tracking>
-          <maximum_range>81.</maximum_range>
-          <minimum_range>0.5</minimum_range>
+          <maximum_range>90.</maximum_range>
+          <minimum_range>0.7</minimum_range>
           <!-- ENU to SFM -->
           <reference_frame>0 0 0 0 0 -1.570796</reference_frame>
         </gz:dvl>
@@ -85,72 +87,27 @@
       <self_collide>0</self_collide>
       <enable_wind>0</enable_wind>
       <kinematic>0</kinematic>
-      <visual name="explorer_base_link_visual">
-        <pose>0 0 0 1.5708 0 1.5708</pose>
+      <visual name="whn_base_link_visual">
+        <pose>0 0 0 0 3.141592653589793 0</pose>
         <geometry>
           <mesh>
             <scale>1 1 1</scale>
-            <uri>model://meshes/teledyne_explorer1000/MarineExplorer.dae</uri>
+            <uri>model://meshes/teledyne_whn/WorkhorseNavigator.dae</uri>
           </mesh>
         </geometry>
         <transparency>0</transparency>
         <cast_shadows>1</cast_shadows>
       </visual>
-      <collision name="explorer_base_link_collision">
+      <collision name="whn_base_link_collision">
         <laser_retro>0</laser_retro>
         <max_contacts>10</max_contacts>
-        <pose>0 0 0 3.14159 -1.57079 0</pose>
+        <pose>0 0 0 0 0 0</pose>
         <geometry>
           <cylinder>
-            <radius>0.062</radius>
-            <length>0.326</length>
+            <radius>0.1126</radius>
+            <length>0.2445</length>
           </cylinder>
         </geometry>
-        <surface>
-          <friction>
-            <ode>
-              <mu>1</mu>
-              <mu2>1</mu2>
-              <fdir1>0 0 0</fdir1>
-              <slip1>0</slip1>
-              <slip2>0</slip2>
-            </ode>
-            <torsional>
-              <coefficient>1</coefficient>
-              <patch_radius>0</patch_radius>
-              <surface_radius>0</surface_radius>
-              <use_patch_radius>1</use_patch_radius>
-              <ode>
-                <slip>0</slip>
-              </ode>
-            </torsional>
-          </friction>
-          <bounce>
-            <restitution_coefficient>0</restitution_coefficient>
-            <threshold>1e+06</threshold>
-          </bounce>
-          <contact>
-            <collide_without_contact>0</collide_without_contact>
-            <collide_without_contact_bitmask>1</collide_without_contact_bitmask>
-            <collide_bitmask>1</collide_bitmask>
-            <ode>
-              <soft_cfm>0</soft_cfm>
-              <soft_erp>0.2</soft_erp>
-              <kp>1e+13</kp>
-              <kd>1</kd>
-              <max_vel>0.01</max_vel>
-              <min_depth>0</min_depth>
-            </ode>
-            <bullet>
-              <split_impulse>1</split_impulse>
-              <split_impulse_penetration_threshold>-0.01</split_impulse_penetration_threshold>
-              <soft_cfm>0</soft_cfm>
-              <soft_erp>0.2</soft_erp>
-              <kp>1e+13</kp>
-              <kd>1</kd>
-            </bullet>
-          </contact>
-        </surface>
       </collision>
     </link>
     <static>0</static>
